backend/requirements.txt
langgraph>=1.0.0
langgraph-checkpoint-postgres>=2.0.0
langchain-openai>=0.3.0
fastapi
uvicorn[standard]
psycopg[binary]>=3.1.0   # psycopg3，不是 psycopg2-binary
requests
gradio>=4.0.0
python-dotenv
